% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_dat_list}
\alias{create_dat_list}
\title{Create Stan data list object}
\usage{
create_dat_list(dt, increase = TRUE)
}
\arguments{
\item{dt}{Modified form of original data}

\item{increase}{TRUE (Default) if increase in outcome is desirable.
Set FALSE if increase in outcome is undesirable.}
}
\value{
Data list object for Stan
}
\description{
Create Stan data list object
}
\keyword{internal}
