% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_spl_mat}
\alias{create_spl_mat}
\title{Create data list splines}
\usage{
create_spl_mat(dl)
}
\arguments{
\item{dl}{Stan data list}
}
\value{
Spline matrix
}
\description{
Create data list splines
}
\keyword{internal}
