% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generic_data_checks}
\alias{generic_data_checks}
\title{Generic data checks}
\usage{
generic_data_checks(data, grouping, condition, outcome, case, time)
}
\arguments{
\item{data}{A dataset, ideally a data.frame.}

\item{grouping}{The name of the grouping variable in the dataset.}

\item{condition}{The level of the grouping variable that identifies the
treatment condition.}

\item{outcome}{The name of the outcome variable.}

\item{case}{The name of the variable that identifies different cases
in the dataset.}

\item{time}{The name of the time variable. This must be a series of
positive whole numbers signifiying the time the outcome was measured.}
}
\value{
Data
}
\description{
Generic data checks
}
\keyword{internal}
