% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generic_numeric_checks}
\alias{generic_numeric_checks}
\title{Generic numeric checks}
\usage{
generic_numeric_checks(dt, outcome)
}
\arguments{
\item{dt}{Modified form of original data}

\item{outcome}{The name of the outcome variable.}
}
\value{
Fail if outcome is not numeric
}
\description{
Generic numeric checks
}
\keyword{internal}
