% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssrhom-package.R
\name{ssrhom-package}
\alias{ssrhom-package}
\alias{ssrhom}
\title{The 'ssrhom' package.}
\description{
The 'ssrhom' package.
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.6. https://mc-stan.org
}
