% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ssrhom_list_stats}
\alias{ssrhom_list_stats}
\title{Function to list out effects computed by package.}
\usage{
ssrhom_list_stats(table = TRUE)
}
\arguments{
\item{table}{If TRUE, report statistics in a table describing
each statistic.
If FALSE, simply return statistics as a list.}
}
\description{
Function to list out effects computed by package.
}
