% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tasky}
\alias{tasky}
\title{Dataset from Tasky et al. (2008)}
\format{
\subsection{\code{tasky}}{

A data frame with 70 rows and 3 colums:
\describe{
\item{person}{The case label}
\item{phase}{Intervention phase where "B" is treatment}
\item{count}{The count of on-task intervals out of 6}
\item{time}{Session number}
}
}
}
\source{
Tasky et al. (2008). Using Choice to Increase On-Task Behavior in
Individuals with Traumatic Brain Injury. Journal of Applied Behavior
Analysis, 41(2), 261–265. \url{https://doi.org/10.1901/jaba.2008.41-261}
}
\usage{
tasky
}
\description{
Dataset from Tasky et al. (2008)
}
\keyword{datasets}
