% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssrhom_get_effect.R
\name{ssrhom_get_effect}
\alias{ssrhom_get_effect}
\title{Report an effect of interest}
\usage{
ssrhom_get_effect(res_obj, stat = "nap", interval = 0.95, return_draws = FALSE)
}
\arguments{
\item{res_obj}{Object returned by main function}

\item{stat}{One of "mean", "median",
"mean-diff", "median-diff",
"log-mean-ratio",
"nap", "tau", "pem", "smd_c", or "smd_p"}

\item{interval}{Some quantile interval between 0 and 1}

\item{return_draws}{If TRUE, do not summarize the posterior samples.
If FALSE, summarize the posterior samples.}
}
\value{
Returns dataset.
}
\description{
Report an effect of interest
}
